fix: assignee mixed not working
