+++ src/pages/EditorPage/EditorPage.jsx
@@ -48,10 +48,10 @@ import { ayonApi } from '/src/services/ayon'
 import { confirmDialog } from 'primereact/confirmdialog'
 import BuildHierarchyButton from '/src/containers/HierarchyBuilder'
 import NewSequence from './NewSequence'
-import useShortcuts from '/src/hooks/useShortcuts'
 import { useGetUsersAssigneeQuery } from '/src/services/user/getUsers'
 import confirmDelete from '/src/helpers/confirmDelete'
 import { useGetProjectAnatomyQuery } from '/src/services/project/getProject'
+import EditorPageShortcuts from './EditorPageShortcuts'
 
 const EditorPage = () => {
   const project = useSelector((state) => state.project)
@@ -1769,34 +1769,18 @@ const EditorPage = () => {
     treeData = loadingTreeData
   }
 
-  const shortcuts = [
-    {
-      key: 'n',
-      action: () => setNewEntity('folder'),
-    },
-    {
-      key: 'm',
-      action: () => setNewEntity('sequence'),
-    },
-    {
-      key: 't',
-      action: () => setNewEntity('task'),
-      disabled: disableAddNew,
-    },
-    {
-      key: 'c',
-      action: (e) => handleToggleFolder(e, true),
-      closest: 'tr.type-folder',
-    },
-  ]
-
-  useShortcuts(shortcuts, [disableAddNew, expandedFolders])
-
   //
   // Render the TreeTable
 
   return (
     <main className="editor-page">
+      <EditorPageShortcuts
+        {...{
+          setNewEntity,
+          disableAddNew,
+          handleToggleFolder,
+        }}
+      />
       {newEntity &&
         (newEntity === 'sequence' ? (
           <NewSequence
