+++ src/pages/EditorPage/EditorPanel.jsx
@@ -196,6 +196,8 @@ const EditorPanel = ({
       },
     }
 
+    console.log('create form')
+
     const type = nodes[nodeIds[0]]?.data?.__entityType
 
     if (type) {
@@ -319,6 +321,8 @@ const EditorPanel = ({
       }
 
       if ((finalValue && finalValue !== nodeValue) || isMultiple) {
+        // if value is undefined, skip
+        if (nodeValue === undefined) continue
         // if type arrays check dif
         if (Array.isArray(finalValue)) {
           // if not different skip
